Introduce Mediator pattern for improved separation of concerns
Added a new `ChampmanReborn.Mediator` project to handle commands and queries, leveraging MediatR. Refactored controllers and moved DTOs to the Application layer, simplifying dependencies and enhancing maintainability.

+++ ChampManReborn.Domain/Entities/Nation.cs
@@ -11,5 +11,5 @@ public class Nation
     public int? CapitalCity { get; set; }
     public short NumberClubs { get; set; }
     public int NumberStaff { get; set; }
-    public short Reputation { get; set; }
+    public int Reputation { get; set; }
 }
